Add wrapper for sbrk() (using a trampoline). We now maintain our own "end of heap" pointer, and return that to the user when user calls sbrk(). This is instead of allowing usage of glibc's "end of heap" pointer and sbrk function, because glibc can get confused on restart when dealing with address randomization.
Also generalized the trampoline installation significantly. Further trampolines
will need to implement "func_trampoline", "func_wrapper", and add an entry to
the "setup_trampolines" function.

As an optimization for sbrk, we could leave the trampoline in permanently, as
we never call _real_sbrk. This is as opposed to uninstalling the trampoline
each time it is hit.


git-svn-id: https://dmtcp.svn.sourceforge.net/svnroot/dmtcp/trunk@816 ba6dc0ae-232f-0410-883f-9ebac64fc565