Execute fork() and exec() wrappers in exclusive mode
fork() and exec() wrappers pass on the state/information about the
current process/program to the to-be-created process/program.

There can be a potential race in the wrappers if this information gets
changed between the point where it was acquired and the point where the
process/program is created. An example of this situation would be a
different thread executing an open() call in parallel creating a
file-descriptor, which is not a part of the information/state gathered
earlier. This can result in unexpected behavior and can cause the
program/process to fail.

This patch fixes this by acquiring the Wrapper-protection-lock in
exclusive mode (write-lock) when executing these wrappers. This
guarantees that no other thread would be executing inside a wrapper that
can change the process state/information.

NOTE:
1. Currently, we do not have WRAPPER_EXECUTION_LOCK/UNLOCK for socket()
family of wrapper. That would be fixed in a later commit.
2. We need to comeup with a strategy for certain blocking system calls
that can change the state of the process (e.g. accept).

git-svn-id: https://dmtcp.svn.sourceforge.net/svnroot/dmtcp/trunk@1337 ba6dc0ae-232f-0410-883f-9ebac64fc565