Allow ckpt if SUSPEND msg received during fork().
In other words, allow a child process to join coordinator before SUSPEND
barrier.

This changeset allows a process (a child process of some process in the
current computation) to join the computation after the SUSPEND message
has been broadcasted by the coordinator and before the computation state
is "SUSPENDED".

This comes handy if the SUSPEND message is sent while a process is
executing the fork() wrapper to create a new child process. This child
process should be considered part of the computation and should be
allowed to participate in the current checkpoint.

Currently, a brand new process that is not a child of any process in the
current computation, is not allowed to join the coordinator; this can be
fixed later, if desired.

NOTE: There is a potential race here: If the parent process complete the
fork() wrapper before the child process has a chance to connect to the
coordinator, the coordinator would have moved to the next barrier,
leader-election, and hence the child process won't be able to join the
computation. This is intended to be fixed in a later checkin.

git-svn-id: https://dmtcp.svn.sourceforge.net/svnroot/dmtcp/trunk@1353 ba6dc0ae-232f-0410-883f-9ebac64fc565